Moved auth controllers to a separate file
