+++ api/routes/auth/controllers/index.js
@@ -1,0 +1,8 @@
+const registerUser = require('./registerUser');
+
+const loginUser = require('./loginUser');
+
+module.exports = {
+  registerUser,
+  loginUser,
+};
