+++ api/routes/auth/utils/index.js
@@ -1,0 +1,10 @@
+const jwt = require('jsonwebtoken');
+
+const signToken = (id) =>
+  jwt.sign({ id }, process.env.JWT_SECRET, {
+    expiresIn: process.env.JWT_EXPIRES_IN,
+  });
+
+module.exports = {
+  signToken,
+};
